+++ public/index.html
@@ -28,16 +28,16 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
+    <header class="container text-white">
+      <h4> Pet Medicine</h4>
+    </header>
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
+    <footer class="container d-flex text-white py-2">
+      <p>All contents &copy; 2021 - petmedicine.com | Rights Reserved. </p>
+      <div class="ml-auto">
+        <a href="#" class="d-block text-white">Terms of use</a>
+        <a href="#" class="d-block text-white"> Privacy policy</a>
+      </div>
+    </footer>
   </body>
 </html>
