+++ src/components/About.js
@@ -1,0 +1,13 @@
+// import React, {Component} from 'react';
+
+// class About extends Component {
+//   render() {
+//     return(
+//       <div>
+//         <h1> About Medi Pet</h1>
+//       </div>
+//     )
+//   }
+// }
+
+// export default About;
