+++ src/components/Appointments.js
@@ -1,0 +1,7 @@
+export function Appointments() {
+  return(
+    <div>
+      <h1>hi</h1>
+    </div>
+  );
+}
