+++ src/components/Contact.js
@@ -1,0 +1,13 @@
+import React, {Component} from 'react';
+
+class Contact extends Component {
+  render() {
+    return(
+      <div>
+        <h1> Contact Us!!</h1>
+      </div>
+    )
+  }
+}
+
+export default Contact;
