+++ src/test/functions.test.js
